+++ src/app/modules/cart/cart.module.ts
@@ -1,15 +1,13 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { CartListComponent } from './cart-list-component/cart-list-component';
-
-
+import { SharedModule } from '../shared/shared.module';
 
 @NgModule({
   declarations: [
     CartListComponent
   ],
   imports: [
-    CommonModule
+    SharedModule,
   ],
   exports: [
     CartListComponent