+++ changelog.md
@@ -8,3 +8,8 @@ Services
 6. created FirstComponent
 7. injected services into FirstComponent
 8. created click directive
+
+Pipes
+1. realized OrderByPipe
+2. used default pieps in CartListComponent such as : async, uppercase, currency
+3. modified cartService to return Observable